Review follow-ups: startup anchor, uint32 sentinels, findings record
- spoof_monitor_2hz / recovery_nav_sim: the initial state is the startup
  anchor (valid from epoch 0), so a latch before the first clean close
  no longer freezes on the spoofed posterior
- STRUCT_SPF / protectedNav: eventAnchorEpoch uint32 everywhere; freeSlot
  keeps one class (Coder)
- run_recovery: missing-anchor sentinel is 0, not NaN
- CST_spfParam: note that MAX_ANCHOR_AGE only guards the startup anchor
- docs/REVIEW_FINDINGS.md: confirmed/refuted/unverified review findings
  with status

Regressions unchanged (fed + legacy checksums, scheduler T1-T6,
test_variable_numMeas).

+++ STRUCT_SPF.m
@@ -108,7 +108,7 @@ classdef STRUCT_SPF
 
             % ---- find the first free slot ----
             slotFound = false;
-            freeSlot = 1;
+            freeSlot = uint8(1);
             windowLength = CST_spfParam.WINDOW_LENGTH;
 
             for wIndx = 1:windowLength
@@ -304,7 +304,7 @@ classdef STRUCT_SPF
                 'revalComputed',         logical(revalComputed), ...
                 'dwellCount',            dwellCount, ...
                 'eventLatched',          logical(eventLatched), ...
-                'eventAnchorEpoch',      eventAnchorEpoch, ...
+                'eventAnchorEpoch',      uint32(eventAnchorEpoch), ...
                 'eventProbationStarted', logical(eventProbationStarted), ...
                 'eventProbationVetoed',  logical(eventProbationVetoed), ...
                 'eventHandback',         logical(eventHandback), ...
@@ -324,7 +324,7 @@ classdef STRUCT_SPF
             revalComputed         = false;
             dwellCount            = 0;
             eventLatched          = false;
-            eventAnchorEpoch      = 0;
+            eventAnchorEpoch      = uint32(0);
             eventProbationStarted = false;
             eventProbationVetoed  = false;
             eventHandback         = false;
